+++ CHANGELOG.md
@@ -1,5 +1,10 @@
 # Changelog
 
+## [1.1.1] - 2024-10-25
+### Added
+- 完善安装依赖部分
+  - create后自动安装依赖
+
 ## [1.1.0] - 2024-10-25
 ### Added
 - 新增一个脚本命令
